# Use single-owner code path to accumulate children when rendering