We added `ps.Map.UnsafeMutableSet()` to optimise the case where you know a NodeSet is not shared with any other nodes.  We can arrange this quite easily for the `Children` of each node in `joinResults`, hence speed up rendering.

I had to adjust the smart/dumb merger benchmarks when they showed an improvement, which is spurious because this code isn't involved in merging.

Benchmarks comparison between first commit and last commit:
```
benchmark                            old ns/op      new ns/op      delta
BenchmarkRenderList-2                575964931      553021897      -3.98%
BenchmarkRenderHosts-2               408721601      362484026      -11.31%
BenchmarkRenderControllers-2         334499307      334047789      -0.13%
BenchmarkRenderPods-2                298092408      289996139      -2.72%
BenchmarkRenderContainers-2          195267791      194515228      -0.39%
BenchmarkRenderProcesses-2           120603030      118073991      -2.10%
BenchmarkRenderProcessNames-2        140712909      137637707      -2.19%

benchmark                            old allocs     new allocs     delta
BenchmarkRenderList-2                803614         699045         -13.01%
BenchmarkRenderHosts-2               603945         508666         -15.78%
BenchmarkRenderControllers-2         427973         388357         -9.26%
BenchmarkRenderPods-2                383876         343816         -10.44%
BenchmarkRenderContainers-2          253831         238770         -5.93%
BenchmarkRenderProcesses-2           119587         104522         -12.60%
BenchmarkRenderProcessNames-2        163859         142304         -13.15%

benchmark                            old bytes     new bytes     delta
BenchmarkRenderList-2                91847940      79224980      -13.74%
BenchmarkRenderHosts-2               65673920      54034450      -17.72%
BenchmarkRenderControllers-2         46943949      41555952      -11.48%
BenchmarkRenderPods-2                41884841      36448283      -12.98%
BenchmarkRenderContainers-2          28795293      26194801      -9.03%
BenchmarkRenderProcesses-2           13985262      11506184      -17.73%
BenchmarkRenderProcessNames-2        17980727      14821811      -17.57%
```